Implement useResponsiveProp Hook to extract values from responsive styles
