+++ src/createRestyleFunction.ts
@@ -1,102 +1,5 @@
-import {
-  ResponsiveValue,
-  BaseTheme,
-  Dimensions,
-  RestyleFunctionContainer,
-  RNStyleProperty,
-} from './types';
-import {getKeys} from './typeHelpers';
-
-type PropValue = string | number | undefined | null;
-
-type StyleTransformFunction<
-  Theme extends BaseTheme,
-  K extends keyof Theme | undefined,
-  TVal
-> = (params: {value: TVal | null; theme: Theme; themeKey?: K}) => TVal | null;
-
-const getValueForScreenSize = <Theme extends BaseTheme, TVal>({
-  responsiveValue,
-  breakpoints,
-  dimensions,
-}: {
-  responsiveValue: {[key in keyof Theme['breakpoints']]?: TVal};
-  breakpoints: Theme['breakpoints'];
-  dimensions: Dimensions;
-}): TVal | null => {
-  const sortedBreakpoints = Object.entries(breakpoints).sort((valA, valB) => {
-    return valA[1] - valB[1];
-  });
-  const {width} = dimensions;
-  return sortedBreakpoints.reduce<TVal | null>(
-    (acc, [breakpoint, minWidth]) => {
-      if (width >= minWidth && responsiveValue[breakpoint] !== undefined)
-        return responsiveValue[breakpoint] as TVal;
-      return acc;
-    },
-    null,
-  );
-};
-
-const isResponsiveObjectValue = <Theme extends BaseTheme, TVal>(
-  val: ResponsiveValue<TVal, Theme>,
-  theme: Theme,
-): val is {[Key in keyof Theme['breakpoints']]?: TVal} => {
-  if (!val) return false;
-  if (typeof val !== 'object') return false;
-  return getKeys(val).reduce((acc: boolean, key) => {
-    return acc && theme.breakpoints[key as string] !== undefined;
-  }, true);
-};
-
-type ValueOf<T> = T[keyof T];
-
-function isThemeKey<Theme extends BaseTheme>(
-  theme: Theme,
-  K: keyof Theme | undefined,
-): K is keyof Theme {
-  return theme[K as keyof Theme];
-}
-
-const getValue = <
-  TVal extends PropValue,
-  Theme extends BaseTheme,
-  K extends keyof Theme | undefined
->(
-  propValue: ResponsiveValue<TVal, Theme>,
-  {
-    theme,
-    transform,
-    dimensions,
-    themeKey,
-  }: {
-    theme: Theme;
-    transform?: StyleTransformFunction<Theme, K, TVal>;
-    dimensions: Dimensions;
-    themeKey?: K;
-  },
-):
-  | TVal
-  | (K extends keyof Theme ? ValueOf<Theme[K]> : never)
-  | null
-  | undefined => {
-  const val = isResponsiveObjectValue(propValue, theme)
-    ? getValueForScreenSize({
-        responsiveValue: propValue,
-        breakpoints: theme.breakpoints,
-        dimensions,
-      })
-    : propValue;
-  if (transform) return transform({value: val, theme, themeKey});
-  if (isThemeKey(theme, themeKey)) {
-    if (val && theme[themeKey][val as string] === undefined)
-      throw new Error(`Value '${val}' does not exist in theme['${themeKey}']`);
-
-    return val ? theme[themeKey][val as string] : val;
-  }
-
-  return val;
-};
+import {BaseTheme, RestyleFunctionContainer, RNStyleProperty} from './types';
+import {getResponsiveValue, StyleTransformFunction} from './responsiveHelpers';
 
 const createRestyleFunction = <
   Theme extends BaseTheme = BaseTheme,
@@ -121,7 +24,7 @@ const createRestyleFunction = <
     themeKey,
     variant: false,
     func: (props, {theme, dimensions}) => {
-      const value = getValue(props[property], {
+      const value = getResponsiveValue(props[property], {
         theme,
         dimensions,
         themeKey,
