+++ src/types.ts
@@ -52,7 +52,10 @@ export type RestyleFunction<
 };
 
 export type RNStyle = ViewStyle | TextStyle | ImageStyle;
+
 export type RNStyleProperty =
   | keyof ViewStyle
   | keyof TextStyle
   | keyof ImageStyle;
+
+export type PropValue = string | number | undefined | null;
